+++ core/src/main/java/io/github/group18/Model/Items/Pickaxe.java
@@ -17,7 +17,7 @@ public class Pickaxe extends Tool implements Name, Price {
 
     public Result use(String direction) {
         Player player = App.getCurrentGame().getPlayers().get(App.getCurrentGame().getIndexPlayerinControl());
-        Cord tileCord = new Cord(player.getX(), player.getY());
+        Cord tileCord = new Cord((int) player.getX(),(int)  player.getY());
         int dir_x = -1;
         int dir_y = -1;
         System.out.println(direction);
