+++ core/src/main/java/io/github/group18/Model/Items/Shear.java
@@ -10,7 +10,7 @@ public class Shear extends Tool implements Name,Price
     protected int price;
     public String use(String direction){
         Player player = App.getCurrentGame().getPlayers().get(App.getCurrentGame().getIndexPlayerinControl());
-        Cord tileCord = new Cord(player.getX(), player.getY());
+        Cord tileCord = new Cord((int) player.getX(), (int) player.getY());
         int dir_x = -1;
         int dir_y = -1;
         switch (direction.toLowerCase()) {
